refactor(core): split storage/bus types into types::storage submodule
Continues the `core/src/types.rs` kitchen-sink split. Moves `StorageConfig`,
`BusKind`, and `BusConfig` out of the flat root namespace into
`assistant_core::types::storage::*`, dropping their entries from the
`lib.rs` re-export block.

Call-site changes (4 files):
- `crates/bus-nats/src/lib.rs`: `BusConfig` now from `types::storage`.
- `crates/web-ui/src/lib.rs`: `BusKind` hoisted to a single
  crate-level `use` at the top; the two inline `use assistant_core::BusKind`
  blocks inside `#[cfg(feature = "nats")]` arms are dropped (the outer use
  already covers them).
- `crates/interface-cli/src/main.rs`: same shape — hoist `BusKind` import,
  drop the two inline `use` statements.
- `crates/interface-cli/src/cmd_doctor.rs`: split the existing combined
  `assistant_core::{...}` use into a `types::storage::BusKind` import plus
  the remaining crate-root imports.

`types.rs` itself adds `use self::storage::{BusConfig, StorageConfig};` so
the `AssistantConfig` struct fields continue to resolve.

Tests:
- `cargo test -p assistant-core --lib` — 183 pass
- `cargo test -p assistant-bus-nats --lib` — 15 pass
- `cargo clippy --workspace -- -D warnings` — green
